+++ fin-stability/src/main/java/com/example/finstability/service/RetirementPlan.java
@@ -1,6 +1,5 @@
 package com.example.finstability.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -9,10 +8,10 @@ import java.util.List;
 @Service
 public class RetirementPlan {
 //    @Autowired  static SaveData data;
-    public static List<String> retirementPlan( double balance ,double interestRate,
-                                       double currentMonthlyExpense ,double inflation ,
-                                       double yearsUntilRetirement,
-                                       double nextTargetYear)
+    public static List<String> retirementPlan(long balance , double interestRate,
+                                              double currentMonthlyExpense , double inflation ,
+                                              double yearsUntilRetirement,
+                                              double nextTargetYear)
     {
         /*
         -calculates required monthly expense by implying inflation to current monthly expense.
